// Client/client/src/Utility/SD.ts
export enum SD_Roles {
    ADMIN = "admin",
    CUTOMER = "customer",
  }
  
  export enum SD_Status {
    PENDING = "Pending",
    CONFIRMED = "Confirmed",
    PACKED = "Packed",
    OUT_FOR_DELIVERY = "Out for Delivery",
    DELIVERED = "Delivered",
    CANCELLED = "Cancelled",
  }  
  
  
  export enum SD_Categories {
    APPETIZER = "Appetizer",
    ENTREE = "Entrée",
    DESSERT = "Dessert",
    BEVERAGES = "Beverages",
  }
  
  export enum SD_SortTypes {
    PRICE_LOW_HIGH = "Price Low - High",
    PRICE_HIGH_LOW = "Price High - Low",
    NAME_A_Z = "Name A - Z",
    NAME_Z_A = "Name Z - A",
  }